[refactor] common package
